ELEC-195: Updated soft timers, tested on x86 and stm32
