+++ libraries/ms-common/src/x86/test_helpers.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-void TEST_soft_timer_set_counter(uint32_t counter_value) {
+void _test_soft_timer_set_counter(uint32_t counter_value) {
   // Not possible on x86 ignore this.
   return;
 }
