+++ libraries/ms-common/test/test_soft_timer.c
@@ -1,238 +1,186 @@
 #include "soft_timer.h"
-
-#include <signal.h>
-#include <stdbool.h>
-#include <stdint.h>
-
-#include "interrupt.h"
-#include "test_helpers.h"
 #include "unity.h"
+#include "test_helpers.h"
+#include "log.h"
 
-// Medium term callback used across tests.
-static volatile bool s_callback_ran = false;
-static volatile uint8_t s_interrupt_id = 255;
-
-static void prv_soft_timer_test_callback(SoftTimerID id, void* context) {
-  s_interrupt_id = id;
-  s_callback_ran = true;
-}
-
-// Short duration callback used in multi timer test.
-static volatile bool s_short_callback_ran = false;
-static volatile uint8_t s_short_interrupt_id = 255;
-
-static void prv_short_soft_timer_test_callback(SoftTimerID id, void* context) {
-  s_short_interrupt_id = id;
-  s_short_callback_ran = true;
-}
-
-// Long duration callback used in the multi timer test.
-static volatile bool s_long_callback_ran = false;
-static volatile uint8_t s_long_interrupt_id = 255;
-
-static void prv_long_soft_timer_test_callback(SoftTimerID id, void* context) {
-  s_long_interrupt_id = id;
-  s_long_callback_ran = true;
+static void prv_timeout_cb(SoftTimerID timer_id, void *context) {
+  SoftTimerID *cb_id = context;
+  *cb_id = timer_id;
 }
 
 void setup_test(void) {
-  // Reset all timers/interrupts and reset the callback used in all tests.
   interrupt_init();
   soft_timer_init();
-  s_short_callback_ran = false;
-  s_callback_ran = false;
-  s_long_callback_ran = false;
+
+  // Purposely start close to the rollover
+  _test_soft_timer_set_counter(UINT32_MAX - 2000);
 }
 
-void teardown_test(void) {}
+void teardown_test(void) { }
 
-// Software timer end to end test.
-void test_soft_timer_end2end(void) {
-  SoftTimerID timer_id = 255;
+void test_soft_timer_basic(void) {
+  volatile SoftTimerID cb_id = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id = SOFT_TIMER_INVALID_TIMER;
 
   TEST_ASSERT_FALSE(soft_timer_inuse());
 
-  // 0.5 us timer.
-  TEST_ASSERT_OK(soft_timer_start(500, prv_soft_timer_test_callback, NULL, &timer_id));
+  StatusCode ret;
+  ret = soft_timer_start_millis(1, prv_timeout_cb, &cb_id, &id);
+  TEST_ASSERT_OK(ret);
 
-  // Assume the timer id changed to something else. (Not a specific number as this would prevent the
-  // test from passing if we ever modified the internals).
-  TEST_ASSERT_NOT_EQUAL(255, timer_id);
-
-  // Validate it hasn't run yet (possibly flaky);
-  TEST_ASSERT_FALSE(s_callback_ran);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id);
   TEST_ASSERT_TRUE(soft_timer_inuse());
 
-  while (!s_callback_ran) {
-  }
+  while (cb_id == SOFT_TIMER_INVALID_TIMER) { }
 
-  // Once run validate everything is as expected.
-  TEST_ASSERT_TRUE(s_callback_ran);
-  TEST_ASSERT_EQUAL(timer_id, s_interrupt_id);
+  TEST_ASSERT_EQUAL(id, cb_id);
+  TEST_ASSERT_FALSE(soft_timer_inuse());
 }
 
-void test_soft_timer_multitimer(void) {
-  SoftTimerID short_timer = 255;
-  SoftTimerID mid_timer = 255;
-  SoftTimerID long_timer = 255;
+void test_soft_timer_preempt(void) {
+  // Begin medium, longer, short, then long - should finish short, medium, long, longer
+  // Tests ordering: new, append to back, insert into front, insert into middle
+  volatile SoftTimerID cb_id_short = SOFT_TIMER_INVALID_TIMER;
+  volatile SoftTimerID cb_id_medium = SOFT_TIMER_INVALID_TIMER;
+  volatile SoftTimerID cb_id_long = SOFT_TIMER_INVALID_TIMER;
+  volatile SoftTimerID cb_id_longer = SOFT_TIMER_INVALID_TIMER;
 
-  // Verify everything is clear.
-  TEST_ASSERT_FALSE(soft_timer_inuse());
-
-  // 1 ms timer starting immediately
-  TEST_ASSERT_OK(soft_timer_start_millis(1, prv_soft_timer_test_callback, NULL, &mid_timer));
-  // 500 us timer premepts the other timer.
-  TEST_ASSERT_OK(soft_timer_start(500, prv_short_soft_timer_test_callback, NULL, &short_timer));
-  // Queued last timer should return after 1 seconds.
-  TEST_ASSERT_OK(soft_timer_start_seconds(1, prv_long_soft_timer_test_callback, NULL, &long_timer));
-
-  while (!s_short_callback_ran) {
-  }
-
-  // Once the short timer ran make sure the medium timer hasn't run (can be flaky for very short
-  // durations).
-  TEST_ASSERT_FALSE(s_callback_ran);
-  TEST_ASSERT_EQUAL(short_timer, s_short_interrupt_id);
-
-  while (!s_callback_ran) {
-  }
-
-  // Once the mid timer ran make sure the longer timer hasn't run (can be flaky for very short
-  // durations).
-  TEST_ASSERT_FALSE(s_long_callback_ran);
-  TEST_ASSERT_EQUAL(mid_timer, s_interrupt_id);
-
-  while (!s_long_callback_ran) {
-  }
-
-  // Verify the long timer ran.
-  TEST_ASSERT_EQUAL(long_timer, s_long_interrupt_id);
-}
-
-void test_soft_timer_multitimer_short(void) {
-  SoftTimerID short_timer = 255;
-  SoftTimerID mid_timer = 255;
-  SoftTimerID long_timer = 255;
-
-  // Verify everything is clear.
-  TEST_ASSERT_FALSE(soft_timer_inuse());
-
-  // 125 us timer starting immediately
-  TEST_ASSERT_OK(soft_timer_start(125, prv_soft_timer_test_callback, NULL, &mid_timer));
-  // 50 us timer premepts prior timer.
-  TEST_ASSERT_OK(soft_timer_start(50, prv_short_soft_timer_test_callback, NULL, &short_timer));
-  // Queued last timer should return after 200 us.
-  TEST_ASSERT_OK(soft_timer_start(200, prv_long_soft_timer_test_callback, NULL, &long_timer));
-
-  while (!s_short_callback_ran) {
-  }
-
-  // Once the short timer ran make sure the medium timer hasn't run (can be flaky for very short
-  // durations).
-  TEST_ASSERT_FALSE(s_callback_ran);
-  TEST_ASSERT_EQUAL(short_timer, s_short_interrupt_id);
-
-  while (!s_callback_ran) {
-  }
-
-  // Once the mid timer ran make sure the longer timer hasn't run (can be flaky for very short
-  // durations).
-  TEST_ASSERT_FALSE(s_long_callback_ran);
-  TEST_ASSERT_EQUAL(mid_timer, s_interrupt_id);
-
-  while (!s_long_callback_ran) {
-  }
-
-  // Verify the long timer ran.
-  TEST_ASSERT_EQUAL(long_timer, s_long_interrupt_id);
-}
-
-// Test for rollover.
-void test_soft_timer_rollover(void) {
-  SoftTimerID timer_id = 255;
+  SoftTimerID id_short = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id_medium = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id_long = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id_longer = SOFT_TIMER_INVALID_TIMER;
 
   TEST_ASSERT_FALSE(soft_timer_inuse());
-  TEST_soft_timer_set_counter(UINT32_MAX - 500);
-  TEST_ASSERT_OK(soft_timer_start(1000, prv_soft_timer_test_callback, NULL, &timer_id));
-  TEST_ASSERT_NOT_EQUAL(255, timer_id);
 
-  // Validate it hasn't run yet (possibly flaky).
-  TEST_ASSERT_FALSE(s_callback_ran);
+  StatusCode ret;
+
+  ret = soft_timer_start_millis(1, prv_timeout_cb, &cb_id_medium, &id_medium);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_medium);
+
+  ret = soft_timer_start_seconds(1, prv_timeout_cb, &cb_id_longer, &id_longer);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_longer);
+
+  ret = soft_timer_start(500, prv_timeout_cb, &cb_id_short, &id_short);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_short);
+
+  ret = soft_timer_start_millis(500, prv_timeout_cb, &cb_id_long, &id_long);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_long);
+
   TEST_ASSERT_TRUE(soft_timer_inuse());
 
-  while (!s_callback_ran) {
-  }
+  while (cb_id_short == SOFT_TIMER_INVALID_TIMER) { }
 
-  // Once run, validated everything is as expected.
-  TEST_ASSERT_TRUE(s_callback_ran);
-  TEST_ASSERT_EQUAL(timer_id, s_interrupt_id);
-}
+  TEST_ASSERT_EQUAL(id_short, cb_id_short);
+  TEST_ASSERT_NOT_EQUAL(id_medium, cb_id_medium);
+  TEST_ASSERT_NOT_EQUAL(id_long, cb_id_long);
+  TEST_ASSERT_NOT_EQUAL(id_longer, cb_id_longer);
 
-// Test cancellation.
-void test_soft_timer_cancel(void) {
-  SoftTimerID timer_id = 255;
+  while (cb_id_medium == SOFT_TIMER_INVALID_TIMER) { }
+
+  TEST_ASSERT_EQUAL(id_medium, cb_id_medium);
+  TEST_ASSERT_NOT_EQUAL(id_long, cb_id_long);
+  TEST_ASSERT_NOT_EQUAL(id_longer, cb_id_longer);
+
+  while (cb_id_long == SOFT_TIMER_INVALID_TIMER) { }
+
+  TEST_ASSERT_EQUAL(id_long, cb_id_long);
+  TEST_ASSERT_NOT_EQUAL(id_longer, cb_id_longer);
+
+  while (cb_id_longer == SOFT_TIMER_INVALID_TIMER) { }
+
+  TEST_ASSERT_EQUAL(id_longer, cb_id_longer);
 
   TEST_ASSERT_FALSE(soft_timer_inuse());
-  TEST_ASSERT_FALSE(soft_timer_cancel(timer_id));
-  TEST_ASSERT_OK(soft_timer_start(1000, prv_long_soft_timer_test_callback, NULL, &timer_id));
-  TEST_ASSERT_OK(soft_timer_start(500, prv_soft_timer_test_callback, NULL, &timer_id));
-  TEST_ASSERT_NOT_EQUAL(255, timer_id);
-
-  // Validate this hasn't run yet (possibly flaky).
-  TEST_ASSERT_FALSE(s_callback_ran);
-  TEST_ASSERT_TRUE(soft_timer_inuse());
-
-  // Check we can cancel.
-  TEST_ASSERT_TRUE(soft_timer_cancel(timer_id));
-
-  while (!s_long_callback_ran) {
-  }
-
-  // Validate it still hasn't run.
-  TEST_ASSERT_FALSE(s_callback_ran);
 }
 
-// Test checking the duration remaining (loosely)
+void test_soft_timer_cancelled_timer(void) {
+  volatile SoftTimerID cb_id_short = SOFT_TIMER_INVALID_TIMER;
+  volatile SoftTimerID cb_id_long = SOFT_TIMER_INVALID_TIMER;
+
+  SoftTimerID id_short = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id_long = SOFT_TIMER_INVALID_TIMER;
+
+  TEST_ASSERT_FALSE(soft_timer_inuse());
+
+  StatusCode ret;
+  ret = soft_timer_start_millis(1, prv_timeout_cb, &cb_id_long, &id_long);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_long);
+
+  ret = soft_timer_start(10, prv_timeout_cb, &cb_id_short, &id_short);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id_short);
+
+  soft_timer_cancel(id_short);
+
+  while (cb_id_long == SOFT_TIMER_INVALID_TIMER) { }
+
+  TEST_ASSERT_EQUAL(id_long, cb_id_long);
+  TEST_ASSERT_NOT_EQUAL(id_short, cb_id_short);
+
+  TEST_ASSERT_FALSE(soft_timer_inuse());
+}
+
 void test_soft_timer_remaining(void) {
-  SoftTimerID timer_id = 255;
+  volatile SoftTimerID cb_id = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID id = SOFT_TIMER_INVALID_TIMER;
 
   TEST_ASSERT_FALSE(soft_timer_inuse());
 
-  // 500 us timer.
-  const uint32_t duration = 500;
-  TEST_ASSERT_OK(soft_timer_start(duration, prv_soft_timer_test_callback, NULL, &timer_id));
+  StatusCode ret;
+  ret = soft_timer_start_millis(1, prv_timeout_cb, &cb_id, &id);
+  TEST_ASSERT_OK(ret);
+  TEST_ASSERT_NOT_EQUAL(SOFT_TIMER_INVALID_TIMER, id);
 
-  // Assume the timer id changed to something else. (Not a specific number as this would prevent the
-  // test from passing if we ever modified the internals).
-  TEST_ASSERT_NOT_EQUAL(255, timer_id);
-  TEST_ASSERT_TRUE(soft_timer_remaining_time(timer_id) < duration);
-
-  // Validate it hasn't run yet (possibly flaky);
-  TEST_ASSERT_FALSE(s_callback_ran);
-  TEST_ASSERT_TRUE(soft_timer_inuse());
-
-  while (!s_callback_ran) {
+  uint32_t prev_time_remaining = soft_timer_remaining_time(id);
+  while (cb_id == SOFT_TIMER_INVALID_TIMER) {
+    uint32_t time_remaining = soft_timer_remaining_time(id);
+    TEST_ASSERT_TRUE(time_remaining <= prev_time_remaining);
+    prev_time_remaining = time_remaining;
   }
 
-  // Once run validate everything is as expected.
-  TEST_ASSERT_TRUE(s_callback_ran);
-  TEST_ASSERT_EQUAL(timer_id, s_interrupt_id);
-  TEST_ASSERT_EQUAL(soft_timer_remaining_time(timer_id), 0);
+  TEST_ASSERT_EQUAL(0, soft_timer_remaining_time(id));
+  TEST_ASSERT_EQUAL(id, cb_id);
+
+  TEST_ASSERT_FALSE(soft_timer_inuse());
 }
 
-// Create too many software timers.
-void test_soft_timer_resource_exhausted(void) {
-  SoftTimerID timer_id = 255;
+void test_soft_timer_exhausted(void) {
+  volatile SoftTimerID cb_ids[SOFT_TIMER_MAX_TIMERS] = { 0 };
+  volatile SoftTimerID cb_id_single = SOFT_TIMER_INVALID_TIMER;
+  SoftTimerID ids[SOFT_TIMER_MAX_TIMERS] = { 0 };
+  SoftTimerID id_single = SOFT_TIMER_INVALID_TIMER;
 
-  // Start a bunch of 10 second timers we will cancel them before they run.
+  StatusCode ret;
   for (int i = 0; i < SOFT_TIMER_MAX_TIMERS; i++) {
-    TEST_ASSERT_OK(soft_timer_start(10000000, prv_soft_timer_test_callback, NULL, &timer_id));
+    cb_ids[i] = SOFT_TIMER_INVALID_TIMER;
+    ret = soft_timer_start_millis(10, prv_timeout_cb, &cb_ids[i], &ids[i]);
+    TEST_ASSERT_OK(ret);
   }
 
-  // Verify we run out of timers.
-  TEST_ASSERT_EQUAL(STATUS_CODE_RESOURCE_EXHAUSTED,
-                    soft_timer_start(10000000, prv_soft_timer_test_callback, NULL, &timer_id));
+  ret = soft_timer_start(1, prv_timeout_cb, &cb_id_single, NULL);
+  TEST_ASSERT_NOT_EQUAL(STATUS_CODE_OK, ret);
 
-  // Reset all timers before exiting to prevent any strange behaviors.
-  soft_timer_init();
+  // Cancel arbitrary timer
+  soft_timer_cancel(ids[3]);
+
+  ret = soft_timer_start_millis(1, prv_timeout_cb, &cb_id_single, &id_single);
+  TEST_ASSERT_OK(ret);
+
+  while (cb_id_single == SOFT_TIMER_INVALID_TIMER) { }
+
+  TEST_ASSERT_EQUAL(id_single, cb_id_single);
+  TEST_ASSERT_NOT_EQUAL(ids[0], cb_ids[0]);
+
+  TEST_ASSERT_TRUE(soft_timer_inuse());
+
+  for (int i = 0; i < SOFT_TIMER_MAX_TIMERS; i++) {
+    soft_timer_cancel(ids[i]);
+  }
+
+  TEST_ASSERT_FALSE(soft_timer_inuse());
 }
