+++ platform/stm32f0xx/platform.mk
@@ -24,7 +24,7 @@ SCRIPT_DIR := $(PLATFORM_DIR)/scripts
 # Build flags for the device
 CDEFINES := USE_STDPERIPH_DRIVER STM32F072
 CFLAGS := -Wall -Werror -g3 -O3 -std=c99 -Wno-unused-variable -pedantic \
-          -ffunction-sections -fdata-sections -fno-builtin -flto \
+          -ffunction-sections -fdata-sections -fno-builtin \
           --specs=nosys.specs --specs=nano.specs \
           $(ARCH_CLAGS) $(addprefix -D,$(CDEFINES))
